requirement.txt
aiofiles==23.2.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.109.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.10
iniconfig==2.1.0
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyPDF2==3.0.1
pytest==8.3.2
pytest-asyncio==0.23.8
python-docx==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.36.3
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.24.0
wheel==0.45.1
